refactor

+++ controllers/threadController.js
@@ -1,4 +1,5 @@
 const moment = require('moment')
+const { ObjectId } = require('mongoose').Types
 const Thread = require('../models/Thread')
 
 const threadController = {}
@@ -29,7 +30,15 @@ threadController.create = async (req, res, next) => {
 threadController.get = async (req, res, next) => {
   try {
     const { id } = req.params
+    if (!ObjectId.isValid(id)) {
+      return next()
+    }
+
     const thread = await Thread.findById(id).populate('user')
+    if (!thread) {
+      return next()
+    }
+
     const isByUser = thread.user.equals(req.user._id)
     return res.render('viewThread', { thread, isByUser, moment })
   } catch (err) {
@@ -40,7 +49,14 @@ threadController.get = async (req, res, next) => {
 threadController.delete = async (req, res, next) => {
   try {
     const { id } = req.params
-    const thread = await Thread.findById(id).populate('user')
+    if (!ObjectId.isValid(id)) {
+      return next()
+    }
+
+    const thread = await Thread.findById(id)
+    if (!thread) {
+      return next()
+    }
     if (thread.user.equals(req.user._id)) {
       await Thread.findByIdAndDelete(id)
     }
@@ -64,9 +80,16 @@ threadController.renderEditPage = async (req, res, next) => {
 threadController.edit = async (req, res, next) => {
   try {
     const { id } = req.params
-    const { title } = req.body
+    if (!ObjectId.isValid(id)) {
+      return next()
+    }
+
     const thread = await Thread.findById(id).populate('user')
+    if (!thread) {
+      return next()
+    }
     if (thread.user.equals(req.user._id)) {
+      const { title } = req.body
       await Thread.findByIdAndUpdate(id, { title })
     }
     return res.redirect(`/view/${id}`)
